Fix PnLSingle.position lossy bug: int → Decimal | None
Wire ``position`` is a Decimal-shaped string (gate >=127). Earlier
ib_async coerced it to ``int`` via ``int(safe_decimal(...) or 0)``,
silently truncating fractional positions for crypto and other
FRACTIONAL_SIZE_SUPPORT (gate 163) instruments. ``Decimal | None``
end-to-end matches the wire and the rest of the v3.0 Decimal-native
domain.

Changes:
- ``PnLSingle.position`` field: int=0 → Decimal | None = None
- ``PnLSingleArgs.pos`` field: int → Decimal | None
- ``createPnLSingleArgs`` drops the int() coercion; safe_decimal flows
 through. Garbage / nan / empty wire input lands as ``None`` instead
 of a truncated zero.
- Binary path ``self.handlers[95]`` wrap converter list
 int → Decimal so the binary decode coerces consistently.
- ``Wrapper.pnlSingle`` signature: ``pos: int`` → ``pos: Decimal | None``.

Tests updated to assert Decimal values and the None-on-malformed
contract.

Public-API breaking change — flag in v3.0 migration guide.

+++ ib_async/decoder.py
@@ -485,7 +485,7 @@ class Decoder:
             93: self.marketRule,
             94: self.wrap("pnl", [int, float, float, float], skip=1),
             95: self.wrap(
-                "pnlSingle", [int, float, float, float, float, float], skip=1
+                "pnlSingle", [Decimal, float, float, float, float, float], skip=1
             ),
             96: self.historicalTicks,
             97: self.historicalTicksBidAsk,
